Refactor: promote panel primitives to src/components/panels
The rest of the site is adopting the Panels language, so the primitives
and the panels footer move out of home/. Home sections stay put.
