+++ src/components/home/panels/panel-contribute.tsx
@@ -1,7 +1,7 @@
 import { CircleDollarSign, Code, Heart, UserPlus } from "lucide-react";
 
-import { Panel, PanelEyebrow, PanelSub, Row, RowArrow, RowList, RowRight } from "./panel";
-import type { ContentRow } from "./types";
+import { Panel, PanelEyebrow, PanelSub, Row, RowArrow, RowList, RowRight } from "@/components/panels/panel";
+import type { ContentRow } from "@/components/panels/types";
 
 const CONTRIBUTE_PATHWAYS: ContentRow[] = [
 	{
