+++ src/components/panels/panel.tsx
