+++ src/components/panels/panels-footer.tsx
