+++ src/components/panels/types.ts
